fix: double-click only if sufficient flags (#113)
There was a bug that when double clicking a field, the perimeter reveal
was triggered, no matter if the surrounding fields show sufficient flags
or not.

This was solved now by adding a check on the surrounding fields before
actually triggering all these fields.

Furthermore all `any` types were avoided by adding more typings to the
code.

Closes #112